app/requirements.txt
Flask
mysql-connector-python
ldap3
